k-mer reseeding - untested

+++ libs/ma/inc/container/svDb.h
@@ -998,7 +998,7 @@ class SV_DB : public Container
                 rJump.iId = pSvJumpTable->xInsertRow(
                     iSvJumpRunId, rJump.iReadId, rJump.from_start( ), rJump.from_end( ), (uint32_t)rJump.uiFrom,
                     (uint32_t)rJump.uiTo, (uint32_t)rJump.uiQueryFrom, (uint32_t)rJump.uiQueryTo,
-                    rJump.uiNumSupportingNt, rJump.bFromForward, rJump.bToForward, rJump.bFromSeedStart );
+                    (uint32_t)rJump.uiNumSupportingNt, rJump.bFromForward, rJump.bToForward, rJump.bFromSeedStart );
             } // method
         }; // class
